# api/openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  version: 1.0.0
servers:
  - url: http://0.0.0.0:8080/api/v1
  - url: http://localhost:8080/api/v1
  - url: http://127.0.0.1:8080/api/v1
paths:
...
  /categories:
    get:
      summary: Get all categories
      operationId: getCategories
      responses:
        "200":
          $ref: "#/components/responses/CategoryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
    post:
      summary: Create a new category
      operationId: createCategory
      requestBody:
        $ref: "#/components/requestBodies/CategoryCreateRequestBody"
      responses:
        "201":
          $ref: "#/components/responses/CategoryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
  /categories/{id}:
    get:
      tags:
        - categories
      summary: Get a category by ID
      operationId: getCategoryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          $ref: "#/components/responses/CategoryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
    patch:
      tags:
        - categories
      summary: Update a category
      operationId: updateCategoryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        $ref: "#/components/requestBodies/CategoryUpdateRequestBody"
      responses:
        "200":
          $ref: "#/components/responses/CategoryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
    delete:
      tags:
        - categories
      summary: Delete a category
      operationId: deleteCategoryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Category deleted
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
  /transactions:
    get:
      tags:
        - transactions
      summary: Get all transactions for the current user
      operationId: getTransactions
      responses:
        "200":
          $ref: "#/components/responses/TransactionResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
    post:
      tags:
        - transactions
      summary: Create a new transaction
      operationId: createTransaction
      requestBody:
        $ref: "#/components/requestBodies/TransactionCreateRequestBody"
      responses:
        "201":
          $ref: "#/components/responses/TransactionResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
  /transactions/{id}:
    get:
      tags:
        - transactions
      summary: Get a transaction by ID
      operationId: getTransactionById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          $ref: "#/components/responses/TransactionResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
    patch:
      tags:
        - transactions
      summary: Update a transaction
      operationId: updateTransactionById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        $ref: "#/components/requestBodies/TransactionUpdateRequestBody"
      responses:
        "200":
          $ref: "#/components/responses/TransactionResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
    delete:
      tags:
        - transactions
      summary: Delete a transaction
      operationId: deleteTransactionById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Transaction deleted
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
  /monthly-summaries:
    get:
      tags:
        - monthly summaries
      summary: Get all monthly summaries for the current user
      operationId: getMonthlySummaries
      responses:
        "200":
          $ref: "#/components/responses/MonthlySummaryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
    post:
      tags:
        - monthly summaries
      summary: Create a new monthly summary
      operationId: createMonthlySummary
      requestBody:
        $ref: "#/components/requestBodies/MonthlySummaryCreateRequestBody"
      responses:
        "201":
          $ref: "#/components/responses/MonthlySummaryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
  /monthly-summaries/{id}:
    get:
      tags:
        - monthly summaries
      summary: Get a monthly summary by ID
      operationId: getMonthlySummaryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          $ref: "#/components/responses/MonthlySummaryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
    patch:
      tags:
        - monthly summaries
      summary: Update a monthly summary by ID
      operationId: updateMonthlySummaryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        $ref: "#/components/requestBodies/MonthlySummaryUpdateRequestBody"
      responses:
        "200":
          $ref: "#/components/responses/MonthlySummaryResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
    delete:
      tags:
        - monthly summaries
      summary: Delete a monthly summary by ID
      operationId: deleteMonthlySummaryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Monthly summary deleted
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
      security:
        - CsrfAuth: []
components:
  securitySchemes:
    CsrfAuth:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN  # カスタムヘッダー名を指定
  schemas:
...
    CategoryRequest:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [income, expense]
      required:
        - id
        - name
        - type
    CategoryCreateRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [income, expense]
        user_id:
          type: integer
      required:
        - name
        - type
        - user_id
    CategoryUpdateRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [income, expense]
      required:
        - name
        - type
    TransactionRequest:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        category_id:
          type: integer
        date:
          type: string
          format: date
        amount:
          type: number
          format: float
        content:
          type: string
      required:
        - id
        - user_id
        - category_id
        - date
        - amount
    TransactionCreateRequest:
      type: object
      properties:
        user_id:
          type: integer
        category_id:
          type: integer
        date:
          type: string
          format: date
        amount:
          type: number
          format: float
        content:
          type: string
      required:
        - user_id
        - category_id
        - date
        - amount
    TransactionUpdateRequest:
      type: object
      properties:
        user_id:
          type: integer
        category_id:
          type: integer
        date:
          type: string
          format: date
        amount:
          type: number
          format: float
        content:
          type: string
      required:
        - user_id
        - category_id
        - date
        - amount
    MonthlySummaryRequest:
      type: object
      properties:
        id:
          type: integer
        year_month:
          type: string
        income:
          type: number
          format: float
        expense:
          type: number
          format: float
        balance:
          type: number
          format: float
      required:
        - id
        - year_month
        - income
        - expense
        - balance
    MonthlySummaryCreateRequest:
      type: object
      properties:
        user_id:
          type: integer
        year_month:
          type: string
        income:
          type: number
          format: float
        expense:
          type: number
          format: float
        balance:
          type: number
          format: float
      required:
        - user_id
        - year_month
        - income
        - expense
        - balance
    MonthlySummaryUpdateRequest:
      type: object
      properties:
        income:
          type: number
          format: float
        expense:
          type: number
          format: float
        balance:
          type: number
          format: float
        year_month:
          type: string
      required:
        - income
        - expense
        - balance
        - year_month
  requestBodies:
...
    CategoryCreateRequestBody:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CategoryCreateRequest"
    CategoryUpdateRequestBody:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CategoryUpdateRequest"
    TransactionCreateRequestBody:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TransactionCreateRequest"
    TransactionUpdateRequestBody:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TransactionUpdateRequest"               
    MonthlySummaryCreateRequestBody:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MonthlySummaryCreateRequest"
    MonthlySummaryUpdateRequestBody:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MonthlySummaryUpdateRequest"             
  responses:            
...
    CategoryResponse:
      description: Category response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CategoryRequest"
    TransactionResponse:
      description: Transaction response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TransactionRequest"
    MonthlySummaryResponse:
      description: Monthly summary response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MonthlySummaryRequest"
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message